Fix: geps were processed out-of-order causing assertfailures
this bug took me way too long to find, but essentailly it goes like this:
GepMap is a std::map<GepInst*, someTuple>
we insert stuff into the gepmap linearly
When we walked over the gepmap, we require the geps to come out of the map linearly too right? as creating a gep where the geptype does not match the pointertype fails with an error
std::map does not guarantee order

Funny enough, it was working because the gep pointers were probably linearly allocated in memory by the frontend, but when I started fucking with fucntion bodies it was failing
It was good practise for me to figure out this bug though!
I fixed it by turning the maps into vectors<> as we dont need fast lookup anyway